rakuten-trading-bot-v2/requirements.txt
fastapi
uvicorn
requests
feedparser
python-dotenv
google-cloud-firestore
google-cloud-storage
# 署名認証用
# PyJWT
# cryptography
